sed: fix endless loop in "b loop" test.
Very few places actually check for errors from emit, and I actually see
the same endless loop from "sed (GNU sed) 4.4" on current Debian, so I'm
not sure this isn't Broken As Designed, but an endless loop spewing
"short write" (or saying nothing, in the case of GNU sed) really doesn't
feel like useful behavior in face of EPIPE, which really isn't going to
fix itself. Certainly not being able to run the sed tests to completion
is pretty annoying --- which is why, unless we remove this test as
invalid, we should probably also add a SKIP_HOST=1 to the "b loop" test.

Note that even with this fix you'll see the error twice:

  sed: short write: Broken pipe
  sed: short write: Broken pipe

Once from the first = command to fail, and then another from the !FLAG(n)
flush of the pattern space.