date: more test cleanup.
Add the SKIP_HOST=1 for the POSIX inputs to -d that coreutils doesn't support.

Fix some comments now Rob's pointed out that the "weird" format was just
POSIX with implicit CCYY or CC. (I was confused because coreutils rejects
them [as it rejects all POSIX input to -d], but busybox does accept them,
but interprets them differently, as explained in the test comments.)

Also rename the tests to make it clearer that these are all POSIX format.