Android moved the scheduler policy functions in Q.
They're forwarded to libprocessgroup, but we may as well go straight to
the source since neither library is in the NDK anyway.

This code is unfortunate because it means that even `toybox true` ends
up pulling in a JSON parser at runtime, because ps might call
get_sched_policy/get_sched_policy_name. I'll experiment with
dlopen-on-demand in portability.c and see whether the savings are
worthwhile, but for now at least use the current library directly so we
can save *one* dlopen!