blkid: adapt tests for util-linux blkid.
util-linux's blkid doesn't support reading from stdin, so move that to
being a special toyonly test rather than the default.

toybox blkid differs from util-linux in that it doesn't seem to find the
LABEL for the checked-in example f2fs file system (the offset is wrong,
but also f2fs uses LE16 strings), nor does it support the SEC_TYPE that
util-linux blkid shows for several of the test file systems.