losetup: minor fixes.
Fix `losetup -f` to not fail with an error.

Add the missing \n for `losetup -f --show FILE`.

Use decimal for the device number, like the desktop losetup.

Switch to the FLAG macro.

Make the tests runnable as tests, and expand coverage a bit. With this
patch, the tests pass both with and without TEST_HOST on the desktop.
Note though that this patch is part of fixing some real-life losetup
issues, not part of the "test cleanup" I'm also looking at. losetup is
low down that list!