killall: better handling of long names.
Change names_to_pid() so that we can actually match shell scripts with
long names (the code to get the shell script's name was correct, but
there was an extra test preventing us from actually comparing it to the
sought name).

In kill.c itself, remove a dead test for -l and switch to the FLAG()
macro.

Also extend the tests to explicitly cover long and short names.