test.tar: fix tar tests on Android.
There's no /etc/group on Android.

Loop devices are in different places, but that test doesn't seem to
offer anything beyond the block/char special device tests later, so
remove that rather than fix it.

Fix the block/char special tests (the expectations were missing, but
they're never normally run on the host because the mknod fails).

Likewise add the missing expectation to the "ownership" test, which
isn't normally run on the host because the chown fails.

The "sparse without overflow" test was broken because of the mksh/bash
disagreement about whether in `VAR=val shell_function` that's a
temporary assignment to VAR or one that outlives the command. Switch SUM
to being a function like LST before it. Also rewrite the TARHD stuff
because Android (a) doesn't have hd and (b) doesn't have process
substitution in its shell. (I wish I'd noticed TARHD earlier --- I'd
been manually sending the tar files to xxd for debugging instead!)

With this patch, all the tar tests now pass for me on both the device
and host.