truncate.test: work around ext extended attributes space usage.
Weaken the test slightly so that the extra 4KiB used for SELinux is
ignored.

Unfortunately all the du tests suffer the same problem, and I'm less
sure what to do there. Mounting a read-only file system without the
extended attributes might be the best idea? But since this test modifies
the file system...