xargs: add --max-args synonym, -o option, and fix -p.
The Linux kernel uses the --max-args synonym for -n.

Barbarians who use vi need xargs' -o to be able to do something like:

  find -name xargs.c | xargs vi      # Sad vi.
  find -name xargs.c | xargs -o vi   # Happy vi.

The -p option needs fixing to read from /dev/tty because stdin is
otherwise occupied in xargs. I think xargs is the only place that needs
this, so it didn't seem sensible to make all callers to yesno() be
specific about which they wanted, hence the new function.

Also remove the documentation for the build-time XARGS_PEDANTIC option
which isn't actually implemented.

Also add a TODO for -P (which is used by at least one script in the
Linux kernel).