vi: small fixes.
I really needed to be able to edit a file on the device, and this was
the nearest thing handy, and it turns out to be more or less usable for
basic editing, so...

Support cursor keys.

Support :q (since there's currently no record of whether the file's
modified or not, :q is the same as :q!).

Add 'A' to insert at end of line.

Add 'n' to find next after '/'.

Fix backspace all the way to get out of ex command mode.

Fix escape sequences to not hard-code assumptions about the terminal's
default background and foreground colors.

Fix 'spesial' typo for 'special', and remove explicit array sizes.