xargs: various fixes.
Don't run the command with no arguments if we run out of input but
have already run the command at least once. The implementation of "run
the command at least once (unless -r was supplied)" wasn't taking into
account whether or not this was our first time round the loop.

Fix the exit value, and the -- already documented but not implemented --
behavior if a child exits with status 255.

Also extend the tests to cover these cases, plus cases I broke while
coming up with the fix. Add more tests to convince myself that we've
correctly interpreted how -s is supposed to behave, and fix the corner
cases at the bottom end of the range.

This fixes some issues we were seeing trying to build the Android SDK
for (and more importantly, on) macOS.