losetup: Fix memory leaks in loopback_setup()
The function loopback_setup() uses xabspath() to get the loopback path.
This function allocates dynamic memory which should be freed by the
function caller.
But there are early return cases where the dynamic memory is not freed.
Besides the special cases of perror_exit(), for which the "early" free
operation is simply used to silence memory analysis tools, the

  if (racy && errno == EBUSY) return 1;

branch may be a real cause of memory leak.

Fix by adding a new free() in the racy+EBUSY branch and anticipating the
existing free().

Signed-off-by: Alessio Balsini <balsini@android.com>