id: support numeric lookup.
Fall back to converting the "name" to an integer and calling getpwuid().
We need to update `username` for the later call to getgrouplist().

Also fix the separator printing logic to avoid a trailing ',' on `id 0`.

Switch to FLAG() and move some declarations down to where they can be
initialized, both for clarity.

Also add simple tests. Sadly, there's no always-present user that is in
multiple groups.