lsattr.test: clean up and fix.
My previous attempt to fix this worked for ext-with-extents on the
desktop, but not for f2fs-with-encryption on cloud Android devices.

This feels quite a bit cleaner, and has the benefit of actually working
everywhere I've tried it.

I've also added perfunctory testing of -p too, which was missing before.