Canonicalize the usage: lines for the POSIX toys.
This only touches 24 of the 68 toys/posix/ files --- the others were
already canonical.

Potentially contentious, so worth stating explicitly, is that there
were 8 matches for "COMMAND..." amongst all existing help output,
with 7 matches for various inconsistent variants involving something
with "ARG", so I resolved that in favor of using "COMMAND..." (which
is also shorter overall, and avoids nested []s).