chattr: fix exit status, redo the tests.
The e2fsprogs chattr(1) returns failure when it fails to do what was
asked of it, and so should we. Fixing this then reveals a lot of issues
with the tests that were being accidentally swept under the carpet.

The bulk of this patch is going through all the tests, removing the
duplicates and making the remaining tests more thorough.

I've tested this on ext4 and f2fs on a variety of 4.x and 5.x kernel
versions (but nothing older). We might need to reduce the list of
attribtues we try to toggle, but the more thorough tests use
well-supported attributes.

I've also fixed the -R test to actually involve a directory hierarchy.