chattr: improve error messages.
When chattr fails in production, it helps to see what it was trying to
do. Reuse the lsattr format but without the '-'s.

Only read the flags if we have any intention of changing them:
`chattr -p 123` has no reason to read the flags.

Only write the flags back if they actually change: `chattr +a`
shouldn't do anything if that flag is already set, for example.

Switch -p and -v to perror_msg() instead of perror_exit() in case
they're used with -R.

(I suspect that the uses of DIRTREE_ABORT are mistakes given -R, but
I'll leave them until I actually hit this.)