vi: don't keep fd open unnecessarily.
As soon as mmap() is done, we can close the fd. xmmap() also will exit
rather than return failure so we can remove that check, and fdlength()
will fall back to lseek() so there's no need to have the fallback in vi
itself.

Spotted because the `TT.fd = 0` in linelist_unload() seemed suspicious;
-1 would have been more natural.