Fix get_block_device_size() for linux
We were incorrectly passing a pointer to a pointer of an
unsigned long long, when we just wanted to pass a pointer
to the unsigned long long.  This is especially bad on 32-bit
systems, where we're then writing a 64-bits into a 32-bit value
within ioctl.

We fix this to pass a pointer to the unsigned long long.

Test: On 32-bit device, no longer see native crash from toybox
Bug: http://b/151311535
Signed-off-by: Elliott Hughes <enh@google.com>