Fix macOS build.
I'd put the rtc_wkalarm struct in the globals to get it zeroed for
free, but since there's no such type on macOS, that breaks the
build. Rather than define a bogus struct in portability.h, I've
gone for making it an explicitly-zeroed local. (And I've sorted the
locals largest-first.)

Note that the use of struct rtc_time as if it was the same as struct
tm in this code follows the existing code in hwclock, but I worry
that we're going to have trouble with that because of the extra
tm_gmtoff and tm_zone fields in struct tm. But that's a worry for
another day...

This patch also removes the CONFIG_CP_PRESERVE from the macos_miniconfig,
since that option was removed a while back.