ls: fix -h with block counts.
The filter() function modifies st_blocks so it's always 1KiB rather than
512B blocks, but the human-readable output was still assuming 512B. This
meant that `ls -sh` was showing figures half the size of `ls -s`, and
that the "total" line with -h was also off by a factor of 2.

No new test, because I don't know how to write one that would work on
all file systems.

Bug: http://b/153383721