cpio: fix tests by removing --trailer.
f2866cf58a07c9d681ecd6695ea969ac70e59a3d fixed the cpio --trailer check
as it claimed to, but doing so changed our behavior (since it used to be
always-on and is now off unless you supply --trailer).

This broke one of the cpio tests.

Since I can't find any other cpio that has a --trailer option (or its
inverse), this patch removes --trailer. This reverts our behavior to
our old behavior (always include the trailer). Judging by the fact that
the cpio tests pass with TEST_HOST=1, this seems to be the expected
behavior.