Teach mkroot to cross compile additional packages, with dropbear as example.
scripts/mkroot.sh CROSS=sh4 LINUX=~/linux dropbear

No, I'm not going down the rathole of adding lots of packages, but this
shows _how_ to do it if you want to. The hooks are there. They don't have
to be in scripts/root, that's just a default search location, you can provide
a path on the command line or have them be in the $PATH.