wget: do not append toybox version at runtime
The sprintf() call, while technically valid (17 bytes fits in an 18
byte allocation) trips Alpine fortify-headers due to checking for
allocations that could potentially overrun.

The call is pointless anyway -- as we are appending a constant to
another constant, it is better to just let the compiler do so and
calculate the size.  This is supported by ISO C89 and later, and
thus any compiler that would be used to compile toybox.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>