getty: ensure utmp is updated.
The key issues here turned out to be that getty is responsible for
creating the file if it doesn't exist, and that the -H flag doesn't
control whether utmp is updated, but whether or not to override the
hostname within the utmp entry.

While I'm here switch to the more modern utx APIs that all the non-pending
parts of toybox use, and remove the duplication.