toybox: fit list of commands in terminal width.
This can still be pretty ragged because it just leaves space for the
longest name at the end of each line rather than measuring the name that
actually comes next, but at least with this change we never over-run.

I noticed this because ConnectBot on my current device gives me a
60-column terminal.

busybox seems to do this, though it seems to actually measure, judging
by how close it gets to the margin. That doesn't seem worth the effort
though?